added optional check for inclusion of gee indicators or not
